float change

+++ src/components/ItemDetails/PercentLabel/PercentLabel.js
@@ -3,8 +3,8 @@ import React, { Fragment } from "react";
 const PercentLabel = props => {
 
     function Percentage(price, oldPrice){
-      var priceNum = parseInt(price);
-      var oldPriceNum = parseInt(oldPrice);
+      var priceNum = parseFloat(price);
+      var oldPriceNum = parseFloat(oldPrice);
        return 100 - (((priceNum * 100)/ oldPriceNum).toFixed())
     }
   return (
